viewport UI controller
